third_party/blink/web_tests/http/tests/navigation/postredirect-frames-goback1-expected.txt

Tests the following sequence of events:

Submit a form via POST to a frame (redirect-to-go-back.pl).
Form page does a 302 to a static page (top-go-back.html), still within the frame.
Static page does a top-level navigation to another static page (go-back.html)
This static page goes back.

We should end up showing the top-level page with the first static page inside the frame.

--------
Frame: 'target-frame'
--------
Target of redirect after POST request.

============== Back Forward List ==============
curr->  http://127.0.0.1:8000/navigation/postredirect-frames-goback1.html
            http://127.0.0.1:8000/navigation/resources/top-go-back.html (in frame "target-frame")
        http://127.0.0.1:8000/navigation/resources/go-back.html
===============================================
